agents/designer/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
langchain-ollama>=0.2
langchain-core>=0.3
httpx>=0.27
pydantic>=2.9
PyYAML>=6.0
python-json-logger>=2.0
